figured out debugging, actually works now

+++ remove.js
@@ -6,10 +6,12 @@ function walkDir(dir, subdir='/') {
     fs.readdir(dir + subdir, (err, files) => {
         if (err) return console.error(err);
         removeIndex(dir, subdir, files);
+        files = files.filter(file => (
+            file !== 'index.html' &&
+            file !== '.html-index' &&
+            file !== '.html-indexed' // legacy
+        ));
         files.forEach((file) => {
-            if (file === 'index.html') return;
-            if (file === '.html-index') return;
-            if (file === '.html-indexed') return;
             fs.stat(dir + subdir + file, (err, stat) => {
                 if (err) return console.error(err);
                 if (stat && stat.isDirectory()) {
@@ -21,20 +23,20 @@ function walkDir(dir, subdir='/') {
 }
 
 function removeIndex(dir, subdir, files) {
-    if (files.includes('.html-index')) {
+    // legacy: clean out any old .html-indexed files
+    const hasHtmlIndex = files.includes('.html-index');
+    const hasHtmlIndexed = files.includes('.html-indexed');
+    if (hasHtmlIndex) removeFile(dir, subdir, '.html-index');
+    if (hasHtmlIndexed) removeFile(dir, subdir, '.html-indexed');
+    if (hasHtmlIndex || hasHtmlIndexed) {
         removeFile(dir, subdir, 'index.html');
-        removeFile(dir, subdir, '.html-index');
-    }
-    // legacy
-    if (files.includes('.html-indexed')) {
-        removeFile(dir, subdir, '.html-indexed');
     }
 }
 
 function removeFile(dir, subdir, file) {
     fs.unlink(dir + subdir + file, err => {
         if (err) return console.error(err);
-        console.log(`-- removed ${file} for ${subdir}`);
+        console.log(` - removed ${file} for ${subdir}`);
     });
 }
